requirements.txt
# PySpark dependencies
pyspark==3.4.0
py4j==0.10.9.7

# Airflow dependencies
apache-airflow==2.7.1
apache-airflow-providers-amazon==8.5.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-http==4.5.1

# Database drivers
psycopg2-binary==2.9.7
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.4.7

# Data processing
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# Configuration and utilities
python-dotenv==1.0.0
pydantic==2.3.0
click==8.1.7
requests==2.31.0

# Logging and monitoring
structlog==23.1.0
prometheus-client==0.17.1

# AWS SDK
boto3==1.28.17
botocore==1.31.17

# Testing utilities (basic)
pytest==7.4.0
